chore: remove unused changes
