+++ mix.exs
@@ -193,6 +193,7 @@ defmodule DataAggregator.MixProject do
       # Ash Framework
       # Revert when https://github.com/ash-project/ash/pull/1419 is merged
       {:ash, "~> 3.3", override: true, github: "hwuethrich/ash", branch: "fix/batch-callbacks"},
+      # {:ash, "~> 3.3", override: true},
       {:ash_json_api, "~> 1.4"},
       {:ash_phoenix, "~> 2.1"},
       {:ash_postgres, "~> 2.1", override: true},
