+++ test/data_aggregator/records/record_test.exs
@@ -62,7 +62,7 @@ defmodule DataAggregator.RecordTest do
 
       record = Ash.load!(record, [:paper_trail_versions, :encoded_record])
 
-      assert length(record.paper_trail_versions) == 2
+      assert length(record.paper_trail_versions) == 1
       assert record.occ_occurrence_id === record.mte_catalog_number
       assert record.oth_basis_of_record === "PreservedSpecimen"
 
@@ -95,7 +95,7 @@ defmodule DataAggregator.RecordTest do
 
       record = Ash.load!(record, [:paper_trail_versions])
 
-      assert length(record.paper_trail_versions) == 3
+      assert length(record.paper_trail_versions) == 2
     end
 
     test "update/2 with invalid data returns error changeset" do
